+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Change Log
 All notable changes to the "nvidia-smi" extension will be documented in this file.
 
+### 0.0.3
+
+- Support recycle, die, clock style.
+
 ### 0.0.2
 
-Support circle indicator.
+- Support circle style.
 
 ## 0.0.1
+
 - Initial release